Fix a very minor compiler warning

+++ src/Misc/Master.cpp
@@ -1028,7 +1028,7 @@ void Master::ShutUp()
         insefx[nefx]->cleanup();
     for(int nefx = 0; nefx < NUM_SYS_EFX; ++nefx)
         sysefx[nefx]->cleanup();
-    for(int i = 0; i < sizeof(activeNotes)/sizeof(activeNotes[0]); ++i)
+    for(int i = 0; i < int(sizeof(activeNotes)/sizeof(activeNotes[0])); ++i)
         activeNotes[i] = 0;
     vuresetpeaks();
     shutup = 0;
